+++ src/main/java/seedu/recipe/logic/parser/EditCommandParser.java
@@ -20,7 +20,7 @@ import seedu.recipe.logic.commands.EditCommand;
 import seedu.recipe.logic.parser.exceptions.ParseException;
 import seedu.recipe.logic.parser.functional.TryUtil;
 import seedu.recipe.logic.util.RecipeDescriptor;
-import seedu.recipe.model.recipe.IngredientBuilder;
+import seedu.recipe.model.recipe.ingredient.IngredientBuilder;
 import seedu.recipe.model.recipe.Step;
 import seedu.recipe.model.tag.Tag;
