+++ src/test/java/seedu/recipe/model/recipe/NameContainsKeywordsPredicateTest.java
@@ -11,6 +11,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.recipe.model.recipe.ingredient.IngredientBuilder;
 import seedu.recipe.model.tag.Tag;
 
 public class NameContainsKeywordsPredicateTest {
