+++ src/test/java/seedu/recipe/testutil/RecipeBuilder.java
@@ -6,7 +6,7 @@ import java.util.Hashtable;
 import java.util.List;
 import java.util.Set;
 
-import seedu.recipe.model.recipe.IngredientBuilder;
+import seedu.recipe.model.recipe.ingredient.IngredientBuilder;
 import seedu.recipe.model.recipe.Name;
 import seedu.recipe.model.recipe.Recipe;
 import seedu.recipe.model.recipe.RecipeDuration;
