+++ src/test/java/seedu/recipe/testutil/TypicalRecipes.java
@@ -8,7 +8,7 @@ import java.util.function.BinaryOperator;
 import java.util.function.Supplier;
 
 import seedu.recipe.model.RecipeBook;
-import seedu.recipe.model.recipe.IngredientBuilder;
+import seedu.recipe.model.recipe.ingredient.IngredientBuilder;
 import seedu.recipe.model.recipe.Name;
 import seedu.recipe.model.recipe.Recipe;
 import seedu.recipe.model.recipe.RecipeDuration;
@@ -16,7 +16,7 @@ import seedu.recipe.model.recipe.RecipePortion;
 import seedu.recipe.model.recipe.Step;
 import seedu.recipe.model.recipe.ingredient.Ingredient;
 import seedu.recipe.model.recipe.ingredient.IngredientQuantifier;
-import seedu.recipe.model.recipe.ingredient.IngredientUtil;
+import seedu.recipe.model.util.IngredientUtil;
 import seedu.recipe.model.tag.Tag;
 
 /**
